Update main.py
Deleted "noise_scale" and replaced it with sqrt(dt). Changed theta value from 0.8 --> 0.7, not much reason for it, but I was comparing against the Wikipedia page and it used 0.7. Rearranged arguments in firstEquation and secondEquation for easier change of y_init, theta, mu, and sigma. Comments added/changed for clarity.